+++ modular_core/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.1.0 - 2022/09/01
+* feat: Make the export of the bind not need exported dependencies.  [790](https://github.com/Flutterando/modular/issues/790)
+
 ## 2.0.3+1 - 2022/06/02
 * feat: Added type propertie in removeBindContext